move services and interfaces to folders
