+++ Complexity/ComplexityWeb/Startup.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.IO;
 using System.Reflection;
-using Application;
 using Application.Info;
+using Application.QuizService;
 using Application.Repositories;
 using Application.Selectors;
+using Application.TaskService;
 using AutoMapper;
 using ComplexityWebApi.DTO;
 using DataBase;
